create command map abstraction
